[GECO-138] : Adding new controller for reprocessing and removing unwanted email config changes
